Add Block as an array wrapper.
